requirements.txt
fastapi==0.115.8
uvicorn[standard]==0.34.0
httpx==0.28.1
redis==5.2.1
beautifulsoup4==4.13.3
lxml==5.3.1
PyMuPDF==1.25.3
python-dotenv==1.0.1
pydantic-settings==2.8.1
